add mare
