+++ Modules/GameOptionsSender/PlayerGameOptionsSender.cs
@@ -6,6 +6,7 @@ using Il2CppSystem.Linq;
 using InnerNet;
 using TOHE.Roles.AddOns.Common;
 using TOHE.Roles.AddOns.Crewmate;
+using TOHE.Roles.AddOns.Impostor;
 using TOHE.Roles.Crewmate;
 using TOHE.Roles.Impostor;
 using TOHE.Roles.Neutral;
